requirements.txt
# Core dependencies
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.0
matplotlib==3.10.0

# LangChain and related
langchain==0.1.12
langchain-openai==0.0.8
langchain-community>=0.0.28,<0.1
openai==1.12.0
chromadb==0.4.24
tiktoken==0.6.0

# Document processing
PyMuPDF==1.24.0
PyYAML==6.0.1

# API and server
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.3
tenacity==8.2.3
jinja2==3.1.3
markdown==3.5.2

# Development tools
pytest==8.0.2
black==24.2.0
ruff==0.3.0 